Text Search C# Application
I have designed an application to search for queries in text files or
word files and also in metadata files (XML files). In the command line,
the file pattern, file path – indicating the files has to be searched in
a specified file path, text queries or/and metadata queries are passed.
• It is the simplest composite text analysis application which helps the
users to search for texts in text files or word files and metadata tag
contents in Xml files.
• Metadata tool has been created, using which Xml files (Metadata files)
are created for each text file.
• The main purpose of this Project is to support keeping track of large
sets of document files.
This Project is done under the guidance of Dr. Jim Fawcett.